fix: consolidate log related types
