+++ packages/rolldown/src/log/logging.ts
@@ -2,9 +2,38 @@ export type LogLevel = 'info' | 'debug' | 'warn';
 export type LogLevelOption = LogLevel | 'silent';
 export type LogLevelWithError = LogLevel | 'error';
 
-// TODO RollupLog Fields
-export type RollupLog = any;
+export interface RollupLog {
+  binding?: string;
+  cause?: unknown;
+  code?: string;
+  exporter?: string;
+  frame?: string;
+  hook?: string;
+  id?: string;
+  ids?: string[];
+  loc?: {
+    column: number;
+    file?: string;
+    line: number;
+  };
+  message: string;
+  meta?: any;
+  names?: string[];
+  plugin?: string;
+  pluginCode?: unknown;
+  pos?: number;
+  reexporter?: string;
+  stack?: string;
+  url?: string;
+}
+
 export type RollupLogWithString = RollupLog | string;
+
+export interface RollupError extends RollupLog {
+  name?: string;
+  stack?: string;
+  watchFiles?: string[];
+}
 
 export type LogOrStringHandler = (
   level: LogLevelWithError,
