+++ packages/rolldown/src/options/normalized-input-options.ts
@@ -1,6 +1,6 @@
 import type { InputOptions } from '..';
 import { BindingNormalizedOptions } from '../binding';
-import type { LogHandler } from '../types/misc';
+import type { LogHandler } from '../log/log-handler';
 
 export interface NormalizedInputOptions {
   input: string[] | Record<string, string>;
