+++ packages/rolldown/src/plugin/index.ts
@@ -7,10 +7,9 @@ import type { BuiltinPlugin } from '../builtin-plugin/constructors';
 import type { DefinedHookNames } from '../constants/plugin';
 import type { DEFINED_HOOK_NAMES } from '../constants/plugin';
 import type { SYMBOL_FOR_RESOLVE_CALLER_THAT_SKIP_SELF } from '../constants/plugin-context';
-import type { LogLevel } from '../log/logging';
+import type { LogLevel, RollupLog } from '../log/logging';
 import type { NormalizedInputOptions } from '../options/normalized-input-options';
 import type { NormalizedOutputOptions } from '../options/normalized-output-options';
-import type { RollupLog } from '../types/misc';
 import type { ModuleInfo } from '../types/module-info';
 import type { OutputBundle } from '../types/output-bundle';
 import type { RenderedChunk } from '../types/rolldown-output';
